@@ -4,7 +4,7 @@ description = "A high-level, massively parallel programming language"
 license = "Apache-2.0"
 version = "0.2.36"
 edition = "2021"
-rust-version = "1.74"
+rust-version = "1.77"
 exclude = ["tests/"]
 
 [lib]
@@ -20,7 +20,8 @@ required-features = ["cli"]
 lto = true
 
 [features]
-default = ["cli"]
+default = ["cli", "manager"]
+manager = ["toml_edit", "git2", "semver"]
 cli = ["dep:clap"]
 
 [dependencies]
@@ -33,8 +34,9 @@ interner = "0.2.1"
 itertools = "0.11.0"
 loaned = "0.1.0"
 stacker = "0.1"
-bpm = { git = "ssh://git@github.com/HigherOrderCO/Bend-Package-Manager.git", features = ["local"] }
-rpassword = "7.3"
+toml_edit = { version = "0.22", optional = true }
+git2 = { version = "0.19", optional = true }
+semver =  { version = "1.0", optional = true }
 
 [dev-dependencies]
 insta = "1.34.0"
